backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]
python-multipart
